LemonSqueezy Integration

+++ app/api/board/route.js
@@ -27,6 +27,13 @@ export async function POST(req) {
 
     const user = await User.findById(session.user.id);
 
+    if (!user.hasAccess) {
+      return NextResponse.json(
+        { error: "Please subscribe to create boards" },
+        { status: 403 }
+      );
+    }
+
     const board = await Board.create({
       userID: user._id,
       name: body.name,
@@ -58,12 +65,19 @@ export async function DELETE(req) {
     }
     await connectMongo();
 
+    const user = await User.findById(session?.user?.id);
+
+    if (!user.hasAccess) {
+      return NextResponse.json(
+        { error: "Please subscribe to delete boards" },
+        { status: 404 }
+      );
+    }
     await Board.deleteOne({
       _id: boardId,
       userID: session?.user?.id,
     });
 
-    const user = await User.findById(session?.user?.id);
     user.boards = user.boards.filter((id) => id.toString() !== boardId);
     await user.save();
